requirements.pyats.txt
pyats
genie
pyats.contrib
flask
pyyaml
requests
textfsm
netmiko